+++ packages/vue/src/toggle-group/index.ts
@@ -1,0 +1,12 @@
+import { Toggle, type ToggleProps } from './toggle'
+import { ToggleGroup as ToggleGroupRoot, type ToggleGroupProps } from './toggle-group'
+import { useToggleGroupContext, type ToggleGroupContext } from './toggle-group-context'
+
+const ToggleGroup = Object.assign(ToggleGroupRoot, {
+  Root: ToggleGroupRoot,
+  Toggle: Toggle,
+})
+
+export { Toggle, ToggleGroup, useToggleGroupContext }
+
+export type { ToggleGroupContext, ToggleGroupProps, ToggleProps }