+++ packages/vue/src/toggle-group/toggle-group-context.ts
@@ -1,0 +1,6 @@
+import { createContext } from '../context'
+import { type UseToggleGroupReturn } from './use-toggle-group'
+export type ToggleGroupContext = UseToggleGroupReturn
+
+export const [ToggleGroupProvider, useToggleGroupContext] =
+  createContext<ToggleGroupContext>('ToggleGroupContext')