+++ packages/vue/src/toggle-group/use-toggle-group.ts
@@ -1,0 +1,40 @@
+import * as toggleGroup from '@zag-js/toggle-group'
+import { normalizeProps, useMachine, type PropTypes } from '@zag-js/vue'
+import { computed, type ComputedRef } from 'vue'
+import { useEnvironmentContext } from '../environment'
+import type { Optional } from '../types'
+import { useId } from '../utils'
+
+export type UseToggleGroupProps = Optional<toggleGroup.Context, 'id'> & {
+  modelValue?: toggleGroup.Context['value']
+}
+export type UseToggleGroupReturn = ComputedRef<toggleGroup.Api<PropTypes>>
+
+export const useToggleGroup = (
+  props: UseToggleGroupProps,
+  emit: CallableFunction,
+): UseToggleGroupReturn => {
+  const getRootNode = useEnvironmentContext()
+  const context = computed(() => {
+    const { modelValue, ...rest } = props
+    return {
+      ...rest,
+      value: modelValue,
+    }
+  })
+
+  const [state, send] = useMachine(
+    toggleGroup.machine({
+      ...context.value,
+      id: context.value.id ?? useId().value,
+      getRootNode,
+      onChange: (details) => {
+        emit('change', details)
+        emit('update:modelValue', details.value)
+      },
+    }),
+    { context },
+  )
+
+  return computed(() => toggleGroup.connect(state.value, send, normalizeProps))
+}