+++ packages/vue/src/toggle/tests/toggle.test.vue
@@ -1,6 +1,0 @@
-<script setup lang="ts">
-import { Toggle } from '..'
-</script>
-<template>
-  <Toggle>B</Toggle>
-</template>